Adding check to after-prepare hook so that it does not run for android builds

+++ src/scripts/after-prepare.js
@@ -4,8 +4,12 @@ const projectMapFolder = '/platforms/ios/NativeScriptMobilePay/';
 module.exports = function (projectData, hookArgs) {
     var fs = require('fs');
     let targetMapFolder = projectData.projectDir + projectMapFolder;
-    if (!fs.existsSync(targetMapFolder)) {
-        fs.mkdirSync(targetMapFolder);
+    if(fs.existsSync('/platforms/ios/')) {
+        if (!fs.existsSync(targetMapFolder)) {
+            fs.mkdirSync(targetMapFolder);
+        }
+        fs.copyFileSync(projectData.projectDir + sourceMapLocation, projectData.projectDir + targetMapLocation);
+    } else {
+        console.log("This is not an ios build so not running after prepare.js for Mobile Pay plugin!")
     }
-    fs.copyFileSync(projectData.projectDir + sourceMapLocation, projectData.projectDir + targetMapLocation);
 }
